feat: only add new files to Xcode Colocated group
This solves issue #14 by removing from the project those
files from an existing Colocated group that do not exist
anymore and are not included in the `colocated_files` array,
keeping in the `colocated_group` the ones that are already linked.
In the next iteration of `colocated_files`, the iteration is
skipped if the file already belong to `colocated_group`,
otherwise it is added.